feat/meeting-participant 모임 참여자수 증가/제한 기능

+++ src/main/java/com/ll/FlexGym/domain/Meeting/repository/MeetingRepository.java
@@ -3,5 +3,5 @@ package com.ll.FlexGym.domain.Meeting.repository;
 import com.ll.FlexGym.domain.Meeting.entity.Meeting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MeetingRepository extends JpaRepository<Meeting, Integer> {
+public interface MeetingRepository extends JpaRepository<Meeting, Long> {
 }
